refactor: Organized things (gen utils, cubes utils, gl-utils.
